refactor: rename app.py > server.py
